requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.12
python-dotenv==1.0.1

# PDF parsing
pdfminer.six==20231228

# Scoring
scikit-learn==1.5.2
sentence-transformers==3.2.1
numpy==1.26.4
rank-bm25==0.2.2
nltk==3.9.1

# LLM judge + keyword extraction
openai==1.70.0
jinja2==3.1.6
pyyaml==6.0.2

# Validation
pydantic==2.9.2

# Evaluation notebook
pandas==2.2.3
matplotlib==3.9.2
scipy==1.14.1
jupyter==1.1.1
ipykernel==6.29.5
requests==2.32.3

# torch is pulled in by sentence-transformers.
# For CPU-only installs run:
#   pip install torch --index-url https://download.pytorch.org/whl/cpu
# before installing this file.
